# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AccessPurchaseInput {
  /** @minLength 3 */
  email: string;
  plan?: string;
}

export interface AccessGrant {
  id: number;
  email: string;
  apiKey: string;
  plan: string;
  amountCents: number;
  createdAt: string;
}

export interface ClassifySampleInput {
  /** @minLength 1 */
  errorMessage: string;
}

export interface ClassificationResult {
  category: string;
  tier: string;
  confidence: number;
  reasoning: string;
}