# lib/api-zod/src/generated/types/accessGrant.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccessGrant {
  id: number;
  email: string;
  apiKey: string;
  plan: string;
  amountCents: number;
  createdAt: Date;
}