# lib/api-zod/src/generated/types/accessPurchaseInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccessPurchaseInput {
  /** @minLength 3 */
  email: string;
  plan?: string;
}